third_party/blink/web_tests/virtual/clear-site-data-client-hints-support/external/wpt/client-hints/clear-site-data/set-client-hints-after-clear-cache.https-expected.txt

This is a testharness.js-based test.
FAIL Setting client hints after Clear-Site-Data for cache shouldn't allow the client hints to take. assert_equals: expected "MissingDeviceMemory" but got "HadDeviceMemory"
Harness: the test ran to completion.
